Improve UserPicker component
based on feedback from Quentin.
