+++ frontend/src/identity/graphql/queries.graphql
@@ -63,9 +63,3 @@ query RegisterPage {
     passwordRequirements
   }
 }
-
-query GetUsers($query: String!, $workspaceSlug: String!) {
-  users(query: $query, workspaceSlug: $workspaceSlug) {
-    ...User_user
-  }
-}
