+++ frontend/src/workspaces/features/UserPicker/UserPicker.graphql
@@ -1,0 +1,5 @@
+query GetUsers($query: String!, $workspaceSlug: String!) {
+  users(query: $query, workspaceSlug: $workspaceSlug) {
+    ...User_user
+  }
+}
